requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
bcrypt==4.0.1
Jinja2==3.1.2
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
